stockaibe/be/pyproject.toml
[tool.poetry]
name = "stockaibe-be"
version = "0.1.0"
description = "Stock crawler limiter admin backend"
authors = ["AI Assistant <assistant@example.com>"]
packages = [{ include = "stockaibe_be", from = "src" }]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.111.0"
uvicorn = { version = "^0.29.0", extras = ["standard"] }
sqlmodel = "^0.0.22"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
apscheduler = "^3.10.4"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
streamlit = "^1.35.0"
pandas = "^2.2.0"
plotly = "^5.20.0"
redis = "^5.0.0"

[tool.poetry.group.dev.dependencies]
hiredis = { version = "^2.3.0", optional = true }
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
